+++ boundary.h
@@ -1,3 +1,3 @@
 #pragma once
 
-void boundarypsi(float *psi, int m, int n, int b, int h, int w);
+void boundarypsi(double *psi, int m, int n, int b, int h, int w);
